+++ recipe/src/recipe.rs
@@ -1,6 +1,6 @@
 use std::{borrow::Cow, env, fs, path::Path};
 
-use anyhow::Result;
+use anyhow::{Context, Result};
 use blue_build_utils::constants::{
     CI_COMMIT_REF_NAME, CI_COMMIT_SHORT_SHA, CI_DEFAULT_BRANCH, CI_MERGE_REQUEST_IID,
     CI_PIPELINE_SOURCE, GITHUB_EVENT_NAME, GITHUB_REF_NAME, GITHUB_SHA, PR_EVENT_NUMBER,
@@ -14,37 +14,93 @@ use typed_builder::TypedBuilder;
 
 use crate::{Module, ModuleExt};
 
+/// The build recipe.
+///
+/// This is the top-level section of a recipe.yml.
+/// This will contain information on the image and its
+/// base image to assist with building the Containerfile
+/// and tagging the image appropriately.
 #[derive(Default, Serialize, Clone, Deserialize, Debug, TypedBuilder)]
 pub struct Recipe<'a> {
+    /// The name of the user's image.
+    ///
+    /// This will be set on the `org.opencontainers.image.title` label.
     #[builder(setter(into))]
     pub name: Cow<'a, str>,
 
+    /// The description of the user's image.
+    ///
+    /// This will be set on the `org.opencontainers.image.description` label.
     #[builder(setter(into))]
     pub description: Cow<'a, str>,
 
+    /// The base image from which to build the user's image.
     #[serde(alias = "base-image")]
     #[builder(setter(into))]
     pub base_image: Cow<'a, str>,
 
+    /// The version/tag of the base image.
     #[serde(alias = "image-version")]
     #[builder(setter(into))]
     pub image_version: Cow<'a, str>,
 
-    #[serde(alias = "blue-build-tag")]
+    /// The version of `bluebuild` to install in the image
+    #[serde(alias = "blue-build-tag", skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(into, strip_option))]
     pub blue_build_tag: Option<Cow<'a, str>>,
 
+    /// Alternate tags to the `latest` tag to add to the image.
+    ///
+    /// If `alt-tags` is not supplied by the user, the build system
+    /// will assume `latest` and will also tag with the
+    /// timestamp with no version (e.g. `20240429`).
+    ///
+    /// Any user input will override the `latest` and timestamp tags.
+    #[serde(alias = "alt-tags", skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(into, strip_option))]
+    pub alt_tags: Option<Vec<Cow<'a, str>>>,
+
+    /// The modules extension of the recipe.
+    ///
+    /// This holds the list of modules to be run on the image.
     #[serde(flatten)]
     pub modules_ext: ModuleExt<'a>,
 
+    /// Extra data that the user might have added. This is
+    /// done in case we serialize the data to a yaml file
+    /// so that we retain any unused information.
     #[serde(flatten)]
     #[builder(setter(into))]
     pub extra: IndexMap<String, Value>,
 }
 
 impl<'a> Recipe<'a> {
+    /// Generate a list of tags based on the OS version.
+    ///
+    /// ## CI
+    /// The tags are generated based on the CI system that
+    /// is detected. The general format for the default branch is:
+    /// - `${os_version}`
+    /// - `${timestamp}-${os_version}`
+    ///
+    /// On a branch:
+    /// - `br-${branch_name}-${os_version}`
+    ///
+    /// In a PR(GitHub)/MR(GitLab)
+    /// - `pr-${pr_event_number}-${os_version}`/`mr-${mr_iid}-${os_version}`
+    ///
+    /// In all above cases the short git sha is also added:
+    /// - `${commit_sha}-${os_version}`
+    ///
+    /// When `alt_tags` are not present, the following tags are added:
+    /// - `latest`
+    /// - `${timestamp}`
+    ///
+    /// ## Locally
+    /// When ran locally, only a local tag is created:
+    /// - `local-${os_version}`
     #[must_use]
-    pub fn generate_tags(&self, os_version: &str) -> Vec<String> {
+    pub fn generate_tags(&self, os_version: u64) -> Vec<String> {
         trace!("Recipe::generate_tags()");
         trace!("Generating image tags for {}", &self.name);
 
@@ -72,8 +128,13 @@ impl<'a> Recipe<'a> {
                 debug!("Running on the default branch");
                 tags.push(os_version.to_string());
                 tags.push(format!("{timestamp}-{os_version}"));
-                tags.push("latest".into());
-                tags.push(timestamp);
+
+                if let Some(alt_tags) = self.alt_tags.as_ref() {
+                    tags.extend(alt_tags.iter().map(ToString::to_string));
+                } else {
+                    tags.push("latest".into());
+                    tags.push(timestamp);
+                }
             } else {
                 debug!("Running on branch {commit_branch}");
                 tags.push(format!("br-{commit_branch}-{os_version}"));
@@ -103,8 +164,13 @@ impl<'a> Recipe<'a> {
             } else if github_ref_name == "live" || github_ref_name == "main" {
                 tags.push(os_version.to_string());
                 tags.push(format!("{timestamp}-{os_version}"));
-                tags.push("latest".into());
-                tags.push(timestamp);
+
+                if let Some(alt_tags) = self.alt_tags.as_ref() {
+                    tags.extend(alt_tags.iter().map(ToString::to_string));
+                } else {
+                    tags.push("latest".into());
+                    tags.push(timestamp);
+                }
             } else {
                 tags.push(format!("br-{github_ref_name}-{os_version}"));
             }
@@ -119,28 +185,29 @@ impl<'a> Recipe<'a> {
         tags.into_iter().map(|t| t.replace('/', "_")).collect()
     }
 
-    /// # Parse a recipe file
-    /// #
+    /// Parse a recipe file
+    ///
     /// # Errors
+    /// Errors when a yaml file cannot be deserialized,
+    /// or a linked module yaml file does not exist.
     pub fn parse<P: AsRef<Path>>(path: &P) -> Result<Self> {
+        trace!("Recipe::parse({})", path.as_ref().display());
+
         let file_path = if Path::new(path.as_ref()).is_absolute() {
             path.as_ref().to_path_buf()
         } else {
             std::env::current_dir()?.join(path.as_ref())
         };
 
-        let recipe_path = fs::canonicalize(file_path)?;
-        let recipe_path_string = recipe_path.display().to_string();
-        debug!("Recipe::parse_recipe({recipe_path_string})");
-
-        let file = fs::read_to_string(recipe_path)?;
+        let file = fs::read_to_string(&file_path)
+            .context(format!("Failed to read {}", file_path.display()))?;
 
         debug!("Recipe contents: {file}");
 
         let mut recipe = serde_yaml::from_str::<Recipe>(&file)
             .map_err(blue_build_utils::serde_yaml_err(&file))?;
 
-        recipe.modules_ext.modules = Module::get_modules(&recipe.modules_ext.modules).into();
+        recipe.modules_ext.modules = Module::get_modules(&recipe.modules_ext.modules)?.into();
 
         Ok(recipe)
     }